feat(GameBoard): replace useDeleteShortcut with useGameBoardShortcuts for improved keyboard handling
